角色分配菜单，添加新的接口 根据模块和语言，取界面Title 组件增加与菜单的关联，并修改接口

+++ src/main/java/com/helioscloud/atlantis/service/MenuService.java
@@ -28,6 +28,7 @@ public class MenuService extends BaseService<MenuMapper, Menu> {
 
     /**
      * 创建菜单
+     * hasChildCatalog 是否有子目录，默认为false,当添加目录时，会把上级目录的该属性设置为true
      *
      * @param menu
      * @return
@@ -49,8 +50,31 @@ public class MenuService extends BaseService<MenuMapper, Menu> {
         if (count != null && count > 0) {
             throw new BizException(RespCode.CODE_NOT_UNION);
         }
-        if(menu.getParentMenuId() == null || "".equals(menu.getParentMenuId())){
+        if (menu.getParentMenuId() == null || "".equals(menu.getParentMenuId()) || 0 == menu.getParentMenuId()) {
             menu.setParentMenuId(0L);//如果没有上级，则默认为0
+            if (1000 == menu.getMenuTypeEnum()) {
+                throw new BizException(RespCode.ROOT_CATALOG_MUST_BE_CATALOG);
+            }
+        } else {
+            //hasChildCatalog 是否有子目录，默认为false,当添加目录时，会把上级目录的该属性设置为true
+            Menu mm = menuMapper.selectById(menu.getParentMenuId());
+            //如果上级是功能，则不允许再添加子功能
+            if (1000 == mm.getMenuTypeEnum()) {
+                throw new BizException(RespCode.MENU_FUNCTION_PARENT_MUST_BE_CATALOG);
+            }
+            // 如果 当前为目录时，需要更新上级目录的hasChildCatalog字段的值
+            if (1001 == menu.getMenuTypeEnum()) {
+                if (mm.getHasChildCatalog() == null || !mm.getHasChildCatalog()) {
+                    mm.setHasChildCatalog(true);
+                    this.updateById(mm);
+                }
+            } else if (1000 == menu.getMenuTypeEnum()) {
+                // 当前为功能时，校验其上级是否有下级目录，如果有，则不允许功能，功能只能添加到最底级的目录、
+                Integer childCatalogCount = menuMapper.selectCount(new EntityWrapper<Menu>().eq("parent_menu_id", mm.getId()).eq("menu_type", 1001));
+                if (childCatalogCount != null && childCatalogCount > 0) {
+                    throw new BizException(RespCode.MENU_PARENT_CATALOG_ERROR);
+                }
+            }
         }
         menuMapper.insert(menu);
         return menu;
@@ -64,6 +88,7 @@ public class MenuService extends BaseService<MenuMapper, Menu> {
      */
     @Transactional
     public Menu updateMenu(Menu menu) {
+        boolean updateMenuType = false;
         //校验
         if (menu == null || menu.getId() == null) {
             throw new BizException(RespCode.ID_NULL);
@@ -82,14 +107,74 @@ public class MenuService extends BaseService<MenuMapper, Menu> {
         if (menu.getIsDeleted() == null || "".equals(menu.getIsDeleted())) {
             menu.setIsDeleted(mm.getIsDeleted());
         }
-        if(menu.getParentMenuId() == null || "".equals(menu.getParentMenuId())){
+        if (menu.getParentMenuId() == null || "".equals(menu.getParentMenuId())) {
             menu.setParentMenuId(mm.getParentMenuId());//如果没有上级，则默认为0
+        }
+        if (menu.getMenuTypeEnum() == null || "".equals(menu.getMenuTypeEnum())) {
+            menu.setMenuTypeEnum(mm.getMenuTypeEnum());
+        } else {
+            //判断是否是 目录 修改为 功能  1000：功能，1001：目录
+            if (1000 == menu.getMenuTypeEnum() && 1001 == mm.getMenuTypeEnum()) {
+                updateMenuType = true;
+            }
+        }
+        //hasChildCatalog 是否有子目录，默认为false,当添加目录时，会把上级目录的该属性设置为true
+        Menu mmm = menuMapper.selectById(menu.getParentMenuId());
+        // 如果 当前为目录时，需要更新上级目录的hasChildCatalog字段的值
+        if (1001 == menu.getMenuTypeEnum()) {
+            if (mmm.getHasChildCatalog() == null || !mmm.getHasChildCatalog()) {
+                mmm.setHasChildCatalog(true);
+                this.updateById(mmm);
+            }
+        } else {
+            // 由目录 修改为 功能 时，判断是否还有除了当前菜单之后的目录，如果不存在，则需要修改hasChildCatalog的值为false
+            if (updateMenuType) {
+                List<Menu> childCatalog = menuMapper.selectList(new EntityWrapper<Menu>()
+                        .eq("parent_menu_id", menu.getParentMenuId())
+                        .eq("menu_type", 1001));
+                if (childCatalog != null && childCatalog.size() == 1) {
+                    if (mmm.getHasChildCatalog() == null || !mmm.getHasChildCatalog()) {
+                        mmm.setHasChildCatalog(true);
+                        this.updateById(mmm);
+                    }
+                }
+            }
         }
         menu.setCreatedBy(mm.getCreatedBy());
         menu.setCreatedDate(mm.getCreatedDate());
         menu.setMenuCode(mm.getMenuCode());
         this.updateById(menu);
+        //20180823 增加校验 禁用上级菜单的时候，把其所有子菜单都禁用掉，但启用的时候，只启用它自己，不递归处理子菜单
+        if (!menu.getIsEnabled().booleanValue() && mm.getIsEnabled().booleanValue()) {
+            updateChildrenMenu(menu);
+        }
         return menu;
+    }
+
+    /**
+     * 禁用父菜单时，递归禁用其所有子菜单
+     * 递归 处理 菜单的子菜单
+     *
+     * @param menu
+     */
+    private void updateChildrenMenu(Menu menu) {
+        //先判断是否有 启用的子菜单
+        List<Menu> list = menuMapper.selectList(new EntityWrapper<Menu>()
+                .eq("is_enabled", true)
+                .eq("parent_menu_id", menu.getId()));
+        if (list != null && list.size() > 0) {
+            //则去禁用子菜单
+            list.stream().forEach(e -> {
+                e.setIsEnabled(false);
+            });
+            //批量保存
+            this.updateBatchById(list);
+            //递归再查子菜单
+            list.stream().forEach(e -> {
+                updateChildrenMenu(e);
+            });
+        }
+        return;
     }
 
     /**
@@ -110,11 +195,15 @@ public class MenuService extends BaseService<MenuMapper, Menu> {
     @Transactional
     public void deleteMenu(Long id) {
         if (id != null) {
+            //如果有启用的子菜单，则不允许删除
+            Integer childCount = menuMapper.selectCount(new EntityWrapper<Menu>()
+                    .eq("is_enabled", true)
+                    .eq("parent_menu_id", id));
+            if (childCount != null && childCount > 0) {
+                throw new BizException(RespCode.HAVING_CHILD_MENU);
+            }
             this.deleteById(id);
         }
-        /*Menu menu = menuMapper.selectById(id);
-        menu.setIsDeleted(true);
-        menuMapper.updateById(menu);*/
     }
 
     /**
@@ -125,17 +214,6 @@ public class MenuService extends BaseService<MenuMapper, Menu> {
     public void deleteBatchMenu(List<Long> ids) {
         if (ids != null && CollectionUtils.isNotEmpty(ids)) {
             this.deleteBatchIds(ids);
-            /*List<Menu> result = null;
-            List<Menu> list = menuMapper.selectBatchIds(ids);
-            if (list != null && list.size() > 0) {
-                result = list.stream().map(menu -> {
-                    menu.setIsDeleted(true);
-                    return menu;
-                }).collect(Collectors.toList());
-            }
-            if(result != null){
-                this.updateBatchById(result);
-            }*/
         }
     }
 
@@ -196,4 +274,5 @@ public class MenuService extends BaseService<MenuMapper, Menu> {
     public Menu getMenuById(Long id) {
         return menuMapper.selectById(id);
     }
+
 }
